logmein_app/requirements-test.txt
unittest2==0.5.1
nose==1.3.0
model-mommy==1.2
django-nose==1.2
